updated langCheckbox in Header
